+++ Godeps/_workspace/src/github.com/cloudfoundry/gosigar/sigar_windows_test.go
@@ -1,0 +1,32 @@
+package sigar_test
+
+import (
+	"os"
+
+	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/gomega"
+
+	sigar "github.com/cloudfoundry/gosigar"
+)
+
+var _ = Describe("SigarWindows", func() {
+	Describe("Memory", func() {
+		It("gets the total memory", func() {
+			mem := sigar.Mem{}
+			err := mem.Get()
+
+			Ω(err).ShouldNot(HaveOccurred())
+			Ω(mem.Total).Should(BeNumerically(">", 0))
+		})
+	})
+
+	Describe("Disk", func() {
+		It("gets the total disk space", func() {
+			usage := sigar.FileSystemUsage{}
+			err := usage.Get(os.TempDir())
+
+			Ω(err).ShouldNot(HaveOccurred())
+			Ω(usage.Total).Should(BeNumerically(">", 0))
+		})
+	})
+})