+++ outputs/kafka/kafka.go
@@ -19,10 +19,10 @@ type Kafka struct {
 }
 
 var sampleConfig = `
-    # URLs of kafka brokers
-    brokers = ["localhost:9092"]
-    # Kafka topic for producer messages
-    topic = "telegraf"
+	# URLs of kafka brokers
+	brokers = ["localhost:9092"]
+	# Kafka topic for producer messages
+	topic = "telegraf"
 `
 
 func (k *Kafka) Connect() error {