+++ testutil/accumulator.go
@@ -151,3 +151,14 @@ func (a *Accumulator) HasFloatValue(measurement string) bool {
 
 	return false
 }
+
+// HasMeasurement returns true if the accumulator has a measurement with the
+// given name
+func (a *Accumulator) HasMeasurement(measurement string) bool {
+	for _, p := range a.Points {
+		if p.Measurement == measurement {
+			return true
+		}
+	}
+	return false
+}